requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
html5lib==1.1
requests==2.32.3
python-dateutil==2.9.0.post0
tqdm==4.66.4